+++ hydra-node/src/Hydra/API/Server.hs
@@ -264,6 +264,8 @@ mkTimedServerOutputFromStateEvent event =
     StateChanged.TickObserved{} -> Nothing
     StateChanged.LocalStateCleared{..} -> Just SnapshotSideLoaded{..}
     StateChanged.Checkpoint{state} -> Just $ EventLogRotated state
+    StateChanged.NodeUnsynced -> Just NodeUnsynced
+    StateChanged.NodeSynced -> Just NodeSynced
 
 -- | Projection to obtain the list of pending deposits.
 projectPendingDeposits :: IsTx tx => [TxIdType tx] -> StateChanged.StateChanged tx -> [TxIdType tx]